backend/requirements.txt
annotated-types==0.7.0
anthropic==0.69.0
anyio==4.11.0
bcrypt==5.0.0
blinker==1.9.0
boto3==1.40.42
botocore==1.40.42
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.1.8
cryptography==46.0.2
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.17.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.0
fastapi==0.118.0
Flask==3.1.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
jmespath==1.0.1
MarkupSafe==3.0.3
mysql-connector-python==9.4.0
numpy==2.0.2
opencv-python==4.12.0.88
packaging==25.0
passlib==1.7.4
pdf2image==1.17.0
pillow==11.3.0
pip==21.2.3
postgrest==2.20.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
pydantic-settings==2.11.0
PyJWT==2.10.1
PyMuPDF==1.26.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.3
realtime==2.20.0
requests==2.32.5
rsa==4.9.1
s3transfer==0.14.0
setuptools==57.4.0
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
storage3==2.20.0
StrEnum==0.4.15
supabase==2.20.0
supabase-auth==2.20.0
supabase-functions==2.20.0
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==1.26.20
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
zipp==3.23.0